+++ mmyolo/datasets/__init__.py
@@ -1,7 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .ppyoloe_coco import PPYOLOECocoDataset
 from .transforms import *  # noqa: F401,F403
 from .utils import BatchShapePolicy, yolov5_collate
-from .yolo_coco import PPYOLOECocoDataset, YOLOv5CocoDataset
+from .yolov5_coco import YOLOv5CocoDataset
 
 __all__ = [
     'YOLOv5CocoDataset', 'PPYOLOECocoDataset', 'BatchShapePolicy',
